Restructure server
* Update docker-compose.yml
