+++ server/apps/fieldsight/management/commands/fetch_from_fieldsight.py
