+++ server/apps/geo/management/commands/load_map_areas.py
