+++ server/apps/geo/management/commands/load_sample_maps.py
