+++ server/cronjobs
@@ -1,3 +1,3 @@
 SHELL=/bin/bash
 
-0 * * * * chmod +x /code/fetch_from_fieldsight.sh; /code/fetch_from_fieldsight.sh
+0 * * * * /code/scripts/fetch_from_fieldsight.sh
