--- server/docker-compose.yml
@@ -1,11 +1,0 @@
-version: '3.2'
-
-services:
-  server:
-    image: devtc/pin-dashboard-server:develop
-    build: .
-    command: bash -c '/code/scripts/run_develop.sh'
-    volumes:
-      - .:/code
-    ports:
-      - '8000:8000'
