+++ server/manage.py
@@ -3,9 +3,10 @@ import os
 import sys
 
 if __name__ == "__main__":
-    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "dashboard.settings")
+    os.environ.setdefault("DJANGO_SETTINGS_MODULE", "pin.settings")
     try:
         from django.core.management import execute_from_command_line
+        from django.conf import settings
     except ImportError as exc:
         raise ImportError(
             "Couldn't import Django. Are you sure it's installed and "
@@ -16,4 +17,5 @@ if __name__ == "__main__":
     if argv[:-1] == 'shell':
         argv.append('-i')
         argv.append('ipython')
+    sys.path.append(settings.APPS_DIR)
     execute_from_command_line(argv)
