+++ server/pin/settings.py
@@ -1,5 +1,5 @@
 """
-Django settings for dashboard project.
+Django settings for pin-dashboard project.
 
 Generated by 'django-admin startproject' using Django 2.0.5.
 
@@ -14,6 +14,7 @@ import os
 
 # Build paths inside the project like this: os.path.join(BASE_DIR, ...)
 BASE_DIR = os.path.dirname(os.path.dirname(os.path.abspath(__file__)))
+APPS_DIR = os.path.join(BASE_DIR, 'apps')
 
 
 # Quick-start development settings - unsuitable for production
@@ -32,6 +33,11 @@ CORS_ORIGIN_ALLOW_ALL = True
 
 # Application definition
 
+LOCAL_APPS = [
+    'geo',
+    'metadata',
+    'fieldsight',
+]
 INSTALLED_APPS = [
     'django.contrib.admin',
     'django.contrib.auth',
@@ -43,10 +49,12 @@ INSTALLED_APPS = [
     'corsheaders',
     'djangorestframework_camel_case',
     'rest_framework',
-
-    'geo',
-    'metadata',
-    'fieldsight',
+] + [
+    '{}.{}.apps.{}Config'.format(
+        APPS_DIR.split('/')[-1],
+        app,
+        ''.join([word.title() for word in app.split('_')]),
+    ) for app in LOCAL_APPS
 ]
 
 MIDDLEWARE = [
@@ -60,12 +68,12 @@ MIDDLEWARE = [
     'django.middleware.clickjacking.XFrameOptionsMiddleware',
 ]
 
-ROOT_URLCONF = 'dashboard.urls'
+ROOT_URLCONF = 'pin.urls'
 
 TEMPLATES = [
     {
         'BACKEND': 'django.template.backends.django.DjangoTemplates',
-        'DIRS': [],
+        'DIRS': [os.path.join(APPS_DIR, 'templates')],
         'APP_DIRS': True,
         'OPTIONS': {
             'context_processors': [
@@ -78,7 +86,7 @@ TEMPLATES = [
     },
 ]
 
-WSGI_APPLICATION = 'dashboard.wsgi.application'
+WSGI_APPLICATION = 'pin.wsgi.application'
 
 
 # Database
@@ -126,7 +134,7 @@ REST_FRAMEWORK = {
         'rest_framework.parsers.MultiPartParser',
     ),
 
-    'EXCEPTION_HANDLER': 'dashboard.exception_handler.custom_exception_handler',
+    'EXCEPTION_HANDLER': 'pin.exception_handler.custom_exception_handler',
     'DEFAULT_PAGINATION_CLASS':
         'rest_framework.pagination.LimitOffsetPagination',
     'PAGE_SIZE': 10000,
@@ -158,7 +166,7 @@ USE_TZ = True
 
 STATIC_URL = '/static/'
 STATICFILES_DIRS = [
-    os.path.join(BASE_DIR, "static"),
+    os.path.join(APPS_DIR, "static"),
 ]
 
 MEDIA_URL = '/media/'
