+++ server/scripts/fetch_from_fieldsight.sh
