+++ server/scripts/run_develop.sh
@@ -7,4 +7,4 @@ crond -L /tmp/cron.logs
 pip3 install -r requirements.txt
 python3 manage.py migrate --no-input
 python3 manage.py createinitialrevisions
-python3 manage.py runserver 0.0.0.0:8000
+python3 manage.py runserver 0.0.0.0:8050
